network: refactor generic socket code
Move generic socket code into separate files. Refactor that prepares
move of socket apis from addresses to the socket class.
