+++ source/common/network/socket_impl.h
@@ -1,0 +1,51 @@
+#pragma once
+
+#include "envoy/network/socket.h"
+
+#include "common/common/assert.h"
+
+namespace Envoy {
+namespace Network {
+
+class SocketImpl : public virtual Socket {
+public:
+  // Network::Socket
+  const Address::InstanceConstSharedPtr& localAddress() const override { return local_address_; }
+  void setLocalAddress(const Address::InstanceConstSharedPtr& local_address) override {
+    local_address_ = local_address;
+  }
+
+  IoHandle& ioHandle() override { return *io_handle_; }
+  const IoHandle& ioHandle() const override { return *io_handle_; }
+  void close() override {
+    if (io_handle_->isOpen()) {
+      io_handle_->close();
+    }
+  }
+  bool isOpen() const override { return io_handle_->isOpen(); }
+  void ensureOptions() {
+    if (!options_) {
+      options_ = std::make_shared<std::vector<OptionConstSharedPtr>>();
+    }
+  }
+  void addOption(const OptionConstSharedPtr& option) override {
+    ensureOptions();
+    options_->emplace_back(std::move(option));
+  }
+  void addOptions(const OptionsSharedPtr& options) override {
+    ensureOptions();
+    Network::Socket::appendOptions(options_, options);
+  }
+  const OptionsSharedPtr& options() const override { return options_; }
+
+protected:
+  SocketImpl(IoHandlePtr&& io_handle, const Address::InstanceConstSharedPtr& local_address)
+      : io_handle_(std::move(io_handle)), local_address_(local_address) {}
+
+  const IoHandlePtr io_handle_;
+  Address::InstanceConstSharedPtr local_address_;
+  OptionsSharedPtr options_;
+};
+
+} // namespace Network
+} // namespace Envoy
